--- client/src/components/Footer.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-class Footer extends Component {
-  render() {
-    return (
-      <div>FOOTER TEST</div>
-    )
-  }
-}
-
-export default Footer
