+++ client/src/components/Links.jsx
@@ -1,25 +1,13 @@
 import React, { Component } from 'react'
 import { Link } from 'react-router-dom'
-import styled from 'styled-components'
-
-const Collapse = styled.div.attrs({
-    className: 'collpase navbar-collapse',
-})``
-
-const List = styled.div.attrs({
-    className: 'navbar-nav mr-auto',
-})``
-
-const Item = styled.div.attrs({
-    className: 'collpase navbar-collapse',
-})``
+import {Collapse, List, Item} from '../styles'
 
 class Links extends Component {
     render() {
         return (
             <React.Fragment>
                 <Link to="/" className="navbar-brand">
-                    My first MERN Application
+                    Chart Diary
                 </Link>
                 <Collapse>
                     <List>
