+++ client/src/components/NaviBar.jsx
@@ -1,18 +1,6 @@
 import React, { Component } from 'react'
-import styled from 'styled-components'
-
-import Logo from './Logo'
 import Links from './Links'
-
-const Container = styled.div.attrs({
-    className: 'container',
-})``
-
-const Nav = styled.nav.attrs({
-    className: 'navbar navbar-expand-lg navbar-dark bg-dark',
-})`
-    margin-bottom: 20 px;
-`
+import {Nav, Container} from '../styles'
 
 class NaviBar extends Component {
     render() {
