+++ client/src/components/index.js
@@ -1,10 +1,8 @@
-import NavBar from './NavBar'
 import NaviBar from './NaviBar'
 import Links from './Links'
-import Logo from './Logo'
 import Entry from './Entry'
 import List from './List'
-import Footer from './Footer'
 import Chart from './Chart'
+import Update from './Update'
 
-export { NavBar, Logo, Entry, List, Footer, NaviBar, Links, Chart }
+export { Entry, List, NaviBar, Links, Chart, Update }
