+++ client/src/pages/ChartPage.jsx
@@ -1,15 +1,7 @@
 import React, { Component } from 'react'
-import ReactTable from 'react-table'
-import api from '../api'
 import {Chart} from '../components'
-
-import styled from 'styled-components'
 import 'react-table/react-table.css'
-
-
-const Wrapper = styled.div`
-    padding: 0 40px 40px 40px;
-`
+import {Wrapper} from '../styles'
 
 class ChartPage extends Component {
     render() {
@@ -20,5 +12,4 @@ class ChartPage extends Component {
       )
     }
 }
-
 export default ChartPage
