+++ client/src/pages/UpdatePage.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from 'react'
+import {Update} from '../components'
+// import 'react-table/react-table.css'
+import {Wrapper} from '../styles'
+
+class UpdatePage extends Component {
+    render() {
+      return (
+        <Wrapper>
+          <Update />
+        </Wrapper>
+      )
+    }
+}
+export default UpdatePage
