+++ client/src/pages/index.js
@@ -1,10 +1,8 @@
-import DiariesList from './DiariesList'
-import DiariesInsert from './DiariesInsert'
-import DiariesUpdate from './DiariesUpdate'
-import DiariesChart from './DiariesChart'
-
 import Home from './Home'
 import ListPage from './ListPage'
 import ChartPage from './ChartPage'
+import EntryPage from './EntryPage'
+import UpdatePage from './UpdatePage'
+import DiariesUpdate from './DiariesUpdate'
 
-export {DiariesInsert, DiariesList, DiariesUpdate, DiariesChart, Home, ListPage, ChartPage}
+export { Home, ListPage, ChartPage, EntryPage, UpdatePage, DiariesUpdate }
