+++ client/src/styles/Button.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const Button = styled.button.attrs({
+  className: `btn btn-primary`,
+})`
+  margin: 15px 15px 15px 5px;
+`
+export default Button
