+++ client/src/styles/CancelButton.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const CancelButton = styled.a.attrs({
+  className: `btn btn-danger`,
+})`
+  margin: 15px 15px 15px 5px;
+`
+
+export default CancelButton
