+++ client/src/styles/Collapse.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Collapse = styled.div.attrs({
+  className: 'collpase navbar-collapse',
+})``
+
+export default Collapse
