+++ client/src/styles/Container.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Container = styled.div.attrs({
+  className: 'container',
+})``
+
+export default Container
