+++ client/src/styles/Delete.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Delete = styled.div`
+color: #ff0000;
+cursor: pointer;
+`
+export default Delete
