+++ client/src/styles/InputText.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const InputText = styled.input.attrs({
+  className: 'form-control',
+})`
+  margin: 5px;
+`
+export default InputText
