+++ client/src/styles/Item.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Item = styled.div.attrs({
+  className: 'collpase navbar-collapse',
+})``
+
+export default Item
