+++ client/src/styles/Label.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+const Label = styled.label`
+    margin: 5px;
+`
+export default Label
