+++ client/src/styles/List.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const List = styled.div.attrs({
+  className: 'navbar-nav mr-auto',
+})``
+
+export default List
