+++ client/src/styles/Nav.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const Nav = styled.nav.attrs({
+  className: 'navbar navbar-expand-lg navbar-dark bg-primary',
+})`
+  margin-bottom: 20 px;
+`
+export default Nav
