+++ client/src/styles/SubTitle.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+
+const SubTitle = styled.h6.attrs({
+  className: 'h6',
+})``
+export default SubTitle
