+++ client/src/styles/Title.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Title = styled.h4.attrs({
+  className: 'h4',
+})``
+
+export default Title
