+++ client/src/styles/Update.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Update = styled.div`
+color: #ef9b0f;
+cursor: pointer;
+`
+export default Update
