+++ client/src/styles/Wrapper.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+    padding: 0 40px 40px 40px;
+`
+export default Wrapper
