+++ client/src/styles/WrapperEntry.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const WrapperEntry = styled.div.attrs({
+  className: 'form-group',
+})`
+  margin: 0 30px;
+`
+export default WrapperEntry
