+++ client/src/styles/index.js
@@ -1,0 +1,18 @@
+import Nav from './Nav'
+import Container from './Container'
+import Wrapper from './Wrapper'
+import WrapperEntry from './WrapperEntry'
+import Title from './Title'
+import SubTitle from './SubTitle'
+import Update from './Update'
+import Delete from './Delete'
+import Label from './Label'
+import InputText from './InputText'
+import Button from './Button'
+import CancelButton from './CancelButton'
+import Collapse from './Collapse'
+import List from './List'
+import Item from './Item'
+
+
+export {Nav, Container, Wrapper, WrapperEntry, Title, SubTitle, Update, Delete, Label, InputText, Button, CancelButton, Collapse, List, Item}
