+++ server/controllers/diary-ctrl.js
@@ -81,7 +81,7 @@ updateDiary = async (req, res) => {
             .catch(error => {
                 return res.status(404).json({
                     error,
-                    message: 'Entry not updated!',
+                    message: 'Diary not updated!',
                 })
             })
     })
